chime: Label android.hardware.drm@1.4-service.clearkey
I hwservicemanager: Since android.hardware.drm@1.0::IDrmFactory/clearkey is not registered, trying to start it as a lazy HAL.
I HidlServiceManagement: getService: Trying again for android.hardware.drm@1.0::IDrmFactory/clearkey...
W libc    : Unable to set property "ctl.interface_start" to "android.hardware.drm@1.0::IDrmFactory/clearkey": error code: 0x20
I hwservicemanager: Tried to start android.hardware.drm@1.0::IDrmFactory/clearkey as a lazy service, but was unable to. Usually this happens when a service is not installed, but if the service is intended to be used as a lazy service, then it may be configured incorrectly.

Signed-off-by: Joey Huab <joey@evolution-x.org>
Signed-off-by: frost. (frostg-012) <frost@nusantararom.org>
Change-Id: Iabd35f0fd93e7d6d35369410d6c92a3f20159ec9
Signed-off-by: Raya Mahardika (rktdnt.) <rktdnt@proton.me>